see desc
Add change password by admin editor.
Add batasan role di UserController and PostController.
Add cancel button.
Add alert if error.
Add pagination to post

+++ app/Http/Controllers/PostController.php
@@ -14,26 +14,12 @@ use App\Http\Controllers\Controller;
 class PostController extends Controller
 {
     public function index(){
-        // $post = DB::table('post')
-        // ->join('users', 'post.id_user', '=', 'users.id')
-        // ->select('post.*','users.role')
-        // ->get();
-
-        $post = Post::with('user')->get();
+        $post = Post::with('user')->paginate(5);
 
         return view('post.post', ['post'=>$post]);
     }
 
     public function lihat_post($id){
-        // $post = DB::table('post')->where('id', $id)->first();
-        // $author = DB::table('users')->where('id', $post->id_user)->first();
-        // $post_tag = DB::table('post_tag')
-        // ->join('tag', 'post_tag.id_tag', '=', 'tag.id')
-        // ->where('id_post', $id)->get();
-        // $post_category = DB::table('post_category')
-        // ->join('category', 'post_category.id_category', '=', 'category.id')
-        // ->where('id_post', $id)->get();
-
         $post = Post::with('user')->where('id',$id)->first();
         $post_category = Post::find($id)->category;
         $post_tag = Post::find($id)->tag;
@@ -42,11 +28,16 @@ class PostController extends Controller
     }
 
     public function delete_post($id){
+        $post = DB::table('post')->where('id', $id)->first();
+        if(Auth::user()->role != '3' && Auth::user()->id != $post->id_user){
+            session()->flash('post_red', 'Anda Tidak Diperbolehkan Melakukan Aksi Ini!');
+            return redirect('/post');
+        }
+        
         DB::table('post')->where('id', $id)->delete();
         DB::table('post_tag')->where('id_post', $id)->delete();
         DB::table('post_category')->where('id_post', $id)->delete();
 
         return redirect('/post');
     }
-
 }
